+++ sdk/examples/client/output/all.rs
@@ -55,14 +55,14 @@ async fn main() -> Result<()> {
     //////////////////////////////////
     // create new alias and nft output
     //////////////////////////////////
-    let alias_output_builder = AliasOutputBuilder::new_with_amount(2_000_000, AliasId::null())?
+    let alias_output_builder = AliasOutputBuilder::new_with_amount(2_000_000, AliasId::null())
         .add_feature(SenderFeature::new(address))
         .add_feature(MetadataFeature::new(vec![1, 2, 3])?)
         .add_immutable_feature(IssuerFeature::new(address))
         .add_unlock_condition(StateControllerAddressUnlockCondition::new(address))
         .add_unlock_condition(GovernorAddressUnlockCondition::new(address));
     // address of the owner of the NFT
-    let nft_output_builder = NftOutputBuilder::new_with_amount(1_000_000, NftId::null())?
+    let nft_output_builder = NftOutputBuilder::new_with_amount(1_000_000, NftId::null())
         .add_unlock_condition(AddressUnlockCondition::new(address));
     let outputs = vec![
         alias_output_builder
@@ -107,13 +107,13 @@ async fn main() -> Result<()> {
     );
     let token_id = TokenId::from(foundry_id);
 
-    let foundry_output_builder = FoundryOutputBuilder::new_with_amount(1_000_000, 1, token_scheme)?
+    let foundry_output_builder = FoundryOutputBuilder::new_with_amount(1_000_000, 1, token_scheme)
         .add_unlock_condition(ImmutableAliasAddressUnlockCondition::new(AliasAddress::from(alias_id)));
 
     let outputs = vec![
         alias_output_builder
             .clone()
-            .with_amount(1_000_000)?
+            .with_amount(1_000_000)
             .with_alias_id(alias_id)
             .with_state_index(1)
             .with_foundry_counter(1)
@@ -151,11 +151,11 @@ async fn main() -> Result<()> {
     let nft_output_id = get_nft_output_id(block.payload().unwrap())?;
 
     let basic_output_builder =
-        BasicOutputBuilder::new_with_amount(1_000_000)?.add_unlock_condition(AddressUnlockCondition::new(address));
+        BasicOutputBuilder::new_with_amount(1_000_000).add_unlock_condition(AddressUnlockCondition::new(address));
 
     let outputs = vec![
         alias_output_builder
-            .with_amount(1_000_000)?
+            .with_amount(1_000_000)
             .with_alias_id(alias_id)
             .with_state_index(2)
             .with_foundry_counter(1)
@@ -177,7 +177,7 @@ async fn main() -> Result<()> {
         // with storage deposit return
         basic_output_builder
             .clone()
-            .with_amount(234_100)?
+            .with_amount(234_100)
             .add_unlock_condition(StorageDepositReturnUnlockCondition::new(
                 address,
                 234_000,