+++ sdk/examples/wallet/07_send_native_tokens.rs
@@ -72,7 +72,7 @@ async fn main() -> Result<()> {
         let rent_structure = account.client().get_rent_structure().await?;
 
         let outputs = vec![
-            BasicOutputBuilder::new_with_minimum_storage_deposit(rent_structure)?
+            BasicOutputBuilder::new_with_minimum_storage_deposit(rent_structure)
                 .add_unlock_condition(AddressUnlockCondition::new(Address::try_from_bech32(bech32_address)?))
                 .with_native_tokens(vec![NativeToken::new(*token_id, U256::from(10))?])
                 .finish_output(account.client().get_token_supply().await?)?,